TraceEvaluationResult doesn't belong in evaluator schema class (#222)
* traceresult doesn't belong in evaluator schema class

* lint
