+++ src/any_agent/evaluation/__init__.py
@@ -1,4 +1,4 @@
-from .evaluate import evaluate
+from .evaluate import TraceEvaluationResult, evaluate
 from .evaluation_case import EvaluationCase
 
-__all__ = ["EvaluationCase", "evaluate"]
+__all__ = ["EvaluationCase", "TraceEvaluationResult", "evaluate"]
